requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
openai>=1.60.0
anthropic>=0.42.0
httpx>=0.28.0
openpyxl>=3.1.5
python-dotenv>=1.0.1
websockets>=14.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.28.0
